requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0
numpy>=1.24.0
scipy>=1.11.0
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
scispacy>=0.5.3
spacy>=3.7.0
sentence-transformers>=2.2.0
wandb>=0.16.0
googlemaps>=4.10.0
